Fix/premature stock deduction

+++ apps/orders/views.py
@@ -51,10 +51,6 @@ def checkout_view(request):
                         quantity=cart_item.quantity
                     )
 
-                    cart_item.sku.stock_quantity = F('stock_quantity') - cart_item.quantity
-                    cart_item.sku.save()
-                    cart_item.sku.refresh_from_db()
-
                 mp_service = MercadoPagoService()
 
                 preference = mp_service.create_payment_preference(order, cart.items.all())
@@ -108,19 +104,19 @@ def mercado_pago_webhook(request):
                 if order_id:
                     order = Order.objects.filter(id=order_id).first()
 
-                    if order and status == 'approved':
+                    if order and status == 'approved' and order.status != 'PAID':
                         order.status = 'PAID'
                         order.save()
+
+                        for item in order.items.all():
+                            item.sku.stock_quantity = F('stock_quantity') - item.quantity
+                            item.sku.save()
 
                         cart = Cart.objects.filter(user=order.customer).first()
                         if cart:
                             cart.items.all().delete()
 
                         print(f'✅ Pedido {order_id} atualizado para PAGO via webhook!')
-    #         payload = json.loads(request.body)
-
-    #         print('WEBHOOK RECEBIDO DO MERCADO PAGO:')
-    #         print(json.dumps(payload, indent=4))
 
             return JsonResponse({'status': 'sucesso'}, status=200)
